+++ jsonjsdb/example/basic/db/__table__.json.js
@@ -1,4 +1,4 @@
-jsonjs.data['__meta__'] = 
+jsonjs.data['__table__'] = 
 [
   {
     "name":"user",